+++ src/components/checkout-panel/checkout-selector/checkout-selector.less
