+++ src/components/checkout-panel/value-options/value-option.less
